+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/HandledErrorOverrideScenario.swift
@@ -8,8 +8,6 @@ import Bugsnag
 
 /**
  Sends a handled Error to Bugsnag and overrides the exception name + message
- Demonstrates adjusting report depth to exclude common error handling code from grouping
- See: https://docs.bugsnag.com/platforms/ios-objc/reporting-handled-exceptions/#depth
  */
 class HandledErrorOverrideScenario: Scenario {
 
@@ -22,7 +20,6 @@ class HandledErrorOverrideScenario: Scenario {
         Bugsnag.notifyError(error) { report in
             report.errorMessage = "Foo"
             report.errorClass = "Bar"
-            report.depth += 2
             report.metadata["account"] = [
                 "items": [400,200]
             ]