+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/UnhandledInternalNotifyScenario.swift
@@ -19,7 +19,6 @@ import Bugsnag
                 ["method":"baz()", "file":"[native code]"],
                 ["method":"is_done()"]
             ]
-            report.attachCustomStacktrace(frames, withType: "fake")
         }
     }
 }